rm redundant typecast

+++ x.c
@@ -125,7 +125,7 @@ int main(int argc, char *argv[])
 	while(read(0, x, 1) > 0) {
 
 		if (needs_escaping(*x)) {
-			printf("\\x%.2x", (int)*x);
+			printf("\\x%.2x", *x);
 		}
 		else {
 			putchar(*x);
